Add files via upload
This my code for a Loan Default project. The objective is to create a machine learning  model that will help SuperLender identify customers that have a high risk to fail to meet their legal loan obligations It was a part of a data science competition on Zindi.  Please find the details for the project here: https://zindi.africa/competitions/data-science-nigeria-challenge-1-loan-default-prediction